src/ragchat/requirements.txt
aws_lambda_powertools
boto3
botocore
fastjsonschema
injector
pypdfium2
python-docx
numpy
